feat!: Detect possible deadlocks when instantiating a parallel iterator.
Deadlocks can happen if the producer for results doesn't start as there
is no free thread on the rayon pool, and the only way for it to become free
is if the iterator produces results.

We now offer a `busy_timeout` in the relevant variants of the
`Parallelism` enumeration to allow controlling how long we will wait
until we abort with an error.